Changed the structure of the directory again.
The directory will be divided in files and sorts. The files and their
generator will be in the Files/ subdirectory, while all the testing will
be placed inside the Sorts/ subdirectory.

Changed the overall way the FileGenerator does its work too. The source
fileMain.cpp now reads the number of elements to write on the files from
stdin, instead of receiving it as an argument. Another change is that
random files have been changed so that they avoid repeating elements.