Added Graph header and implementation files.
The Graph class is a direct child of the Digraph class, overriding
the way in which edges are added to the graph. Since the Digraph
class uses an array of lists to represent edges, then, the Graph
class must ensure that whenever and edge is added from node x to
node y, then it must also add it from node y to node y. All other
operations are left the same.

The Graph class will also have two algorithms programmed into it:
Prim (already done) and Kruskal (next in line).